refactored inheritance structure
