+++ src/main/java/com/epicquestthegame/model/Game.java
@@ -1,15 +1,15 @@
 package com.epicquestthegame.model;
 
 public class Game {
-    private final NodeHandler startedNode;
+    private final Handler startedNode;
     private final String gamerName;
 
-    public Game(NodeHandler startedNodeHandler, String gamerName) {
+    public Game(Handler startedNodeHandler, String gamerName) {
         this.startedNode = startedNodeHandler;
         this.gamerName = gamerName;
     }
 
-    public NodeHandler getStartedNode() {
+    public Handler getStartedNode() {
         return startedNode;
     }
