+++ src/main/java/com/epicquestthegame/model/Handler.java
@@ -1,0 +1,9 @@
+package com.epicquestthegame.model;
+
+import javax.servlet.http.HttpServletRequest;
+
+public interface Handler {
+    void handle(Node node, HttpServletRequest request);
+    void setNext(Handler handler);
+    Node getThisNode();
+}
