--- src/main/java/com/epicquestthegame/model/nodes/KillTheKingNodeHandler.java
@@ -1,11 +1,0 @@
-package com.epicquestthegame.model.nodes;
-
-import com.epicquestthegame.model.AbstractNodeHandler;
-import com.epicquestthegame.model.Node;
-import com.epicquestthegame.model.NodeHandler;
-
-public class KillTheKingNodeHandler extends AbstractNodeHandler {
-    public KillTheKingNodeHandler(NodeHandler defeatNode) {
-        super(defeatNode, Node.KILL_THE_KING);
-    }
-}
