--- src/main/java/com/epicquestthegame/model/nodes/MakeAGangNodeHandler.java
@@ -1,12 +1,0 @@
-package com.epicquestthegame.model.nodes;
-
-import com.epicquestthegame.model.AbstractNodeHandler;
-import com.epicquestthegame.model.Node;
-import com.epicquestthegame.model.NodeHandler;
-
-public class MakeAGangNodeHandler extends AbstractNodeHandler {
-
-    public MakeAGangNodeHandler(NodeHandler defeatNode) {
-        super(defeatNode, Node.MAKE_A_GANG);
-    }
-}
