--- src/main/java/com/epicquestthegame/model/nodes/PrisonNodeHandler.java
@@ -1,12 +1,0 @@
-package com.epicquestthegame.model.nodes;
-
-import com.epicquestthegame.model.AbstractNodeHandler;
-import com.epicquestthegame.model.Node;
-import com.epicquestthegame.model.NodeHandler;
-
-public class PrisonNodeHandler extends AbstractNodeHandler {
-
-    public PrisonNodeHandler(NodeHandler defeatNode) {
-        super(defeatNode, Node.PRISON);
-    }
-}
